fixed some tests in menu.spec.js
